fix relative import for subfunc
